feat:新增周期性定时任务

+++ iw-packaging-parent/iw-points/src/main/java/com/itwray/iw/points/controller/client/PointsRecordsClientController.java
@@ -32,7 +32,7 @@ public class PointsRecordsClientController {
     @PostMapping("/add")
     @Operation(summary = "新增积分记录")
     public Integer add(@RequestBody @Valid PointsRecordsAddDto dto) {
-        return (Integer) pointsRecordsService.add(dto);
+        return pointsRecordsService.add(dto);
     }
 
 }
